+++ src/main/java/leets/bookmark/domain/searchhistory/application/exception/SearchHistoryNotFoundException.java
@@ -1,9 +1,9 @@
-package leets.bookmark.domain.searchhistory.application.exeception;
+package leets.bookmark.domain.searchhistory.application.exception;
 
 import leets.bookmark.global.common.exception.BusinessException;
 
 public class SearchHistoryNotFoundException extends BusinessException {
     public SearchHistoryNotFoundException() {
-        super(SearchHistoryErrorCode.NOT_FOUND);
+        super(SearchHistoryErrorCode.SEARCH_HISTORY_NOT_FOUND_EXCEPTION);
     }
 }
